Fixed id's being ints instead of longs in db/interfaces. Also made an initial setup for notification dispatch system

+++ app/src/main/java/com/sebastiaan/silos/db/interfaces/productInterface.java
@@ -13,7 +13,7 @@ public interface productInterface extends DbInterface<product> {
     List<product> getAll();
 
     @Query("SELECT * FROM product where productID = :id")
-    product findByID(int id);
+    product findByID(long id);
 
     @Query("SELECT * FROM product where productName LIKE :name")
     List<product> findByName(String name);
